feat(model-providers): extract ScopeChipPicker + add B3.1 spec
ScopeChipPicker is a thin controlled component that renders the chip-based
multi-select over (organization, teams, projects). Extracted from the
model-provider create-drawer so the role-based default model lines and any
future scope-picking surface can consume the same primitive without
inheriting the drawer's form-state machinery.

ModelProviderScopeSection becomes a thin shell wiring the form state to
ScopeChipPicker for the create flow, while keeping its existing read-only
chip display for the edit flow.

Spec specs/model-providers/model-resolver-and-registry.feature is the
B3.1 contract — 17 scenarios covering the resolution walk, error shape,
onboarding seed semantics, B2-column migration, and the read+write compat
layer. Implementation lands in follow-up commits on this branch.
