+++ langwatch/src/components/settings/ModelProviderScopeSection.tsx
@@ -1,14 +1,6 @@
-import {
-  Box,
-  Button,
-  createListCollection,
-  HStack,
-  Text,
-  VStack,
-  Wrap,
-} from "@chakra-ui/react";
+import { Button, HStack, Text, VStack, Wrap } from "@chakra-ui/react";
 import { Building2, Folder, Users } from "lucide-react";
-import { useMemo } from "react";
+
 import type {
   ModelProviderScopeType,
   ScopeSelection,
@@ -16,16 +8,9 @@ import type {
   UseModelProviderFormState,
 } from "../../hooks/useModelProviderForm";
 import type { MaybeStoredModelProvider } from "../../server/modelProviders/registry";
-import { Select } from "../ui/select";
 import { SmallLabel } from "../SmallLabel";
 import { ProviderScopeChips } from "./ProviderScopeChips";
-
-type ScopeOption = {
-  value: string;
-  label: string;
-  scopeType: ModelProviderScopeType;
-  scopeId: string;
-};
+import { ScopeChipPicker } from "./ScopeChipPicker";
 
 const SCOPE_DESCRIPTION_SINGLE: Record<ModelProviderScopeType, string> = {
   PROJECT: "Only this project can use this provider.",
@@ -58,28 +43,21 @@ function summariseSelection(scopes: ScopeSelection[]): string {
   return `Shared across ${parts.join(" + ")}.`;
 }
 
-const ScopeIcon = ({ scopeType }: { scopeType: ModelProviderScopeType }) => {
-  if (scopeType === "ORGANIZATION") return <Building2 size={16} aria-hidden />;
-  if (scopeType === "TEAM") return <Users size={16} aria-hidden />;
-  return <Folder size={16} aria-hidden />;
-};
-
 /**
- * Scope picker for model providers.
+ * Model-provider scope picker.
  *
- * For NEW providers, renders a Chakra multi-select over the organization,
- * every team the caller is a member of, and the projects inside those
- * teams. Users pick one or more scopes; every selected entry becomes a
- * ModelProviderScope row on save. The service runs fail-closed authz per
- * entry — selecting a scope the caller cannot manage rejects the whole
- * write, there is no partial-success path.
+ * For NEW providers, renders quick-add chips at the top ("This project",
+ * "This team", "Organization") followed by the shared `ScopeChipPicker`
+ * for the multi-select case. Each chip replaces the current selection
+ * with exactly that single scope — the most common setup is "one scope
+ * per credential", and the picker stays available below for cross-team
+ * or multi-project setups.
  *
  * For EXISTING providers the section is read-only: scope changes on a
- * persisted credential happen by delete+recreate so we never silently
+ * persisted credential happen by delete + recreate so we never silently
  * re-parent a credential across orgs/teams.
  *
- * For personal-account projects (no org/team context) the section
- * renders nothing.
+ * Personal-account projects (no org/team context) render nothing.
  */
 export function ProviderScopeSection({
   state,
@@ -103,76 +81,11 @@ export function ProviderScopeSection({
   organizationName?: string;
   projectId?: string;
   projectName?: string;
-  /** Teams the caller can pick for TEAM scope. Falls back to [{id:teamId}] when omitted. */
   availableTeams?: Array<{ id: string; name: string }>;
-  /** Projects the caller can pick for PROJECT scope. Falls back to [{id:projectId}]. */
   availableProjects?: Array<{ id: string; name: string; teamId?: string }>;
 }) {
   const isExisting = Boolean(provider.id);
   const hasOrgOrTeam = Boolean(organizationId ?? teamId);
-
-  // Build options from the accessible set. Falls back to the active
-  // context IDs when the caller hasn't provided a full team/project
-  // list — single-scope flows keep working unchanged.
-  const options = useMemo<ScopeOption[]>(() => {
-    const out: ScopeOption[] = [];
-    if (organizationId) {
-      out.push({
-        value: `ORGANIZATION:${organizationId}`,
-        label: organizationName ?? "Organization",
-        scopeType: "ORGANIZATION",
-        scopeId: organizationId,
-      });
-    }
-    const teams =
-      availableTeams && availableTeams.length > 0
-        ? availableTeams
-        : teamId
-          ? [{ id: teamId, name: teamName ?? "Team" }]
-          : [];
-    for (const team of teams) {
-      out.push({
-        value: `TEAM:${team.id}`,
-        label: team.name,
-        scopeType: "TEAM",
-        scopeId: team.id,
-      });
-    }
-    const projects =
-      availableProjects && availableProjects.length > 0
-        ? availableProjects
-        : projectId
-          ? [{ id: projectId, name: projectName ?? "Project" }]
-          : [];
-    for (const project of projects) {
-      out.push({
-        value: `PROJECT:${project.id}`,
-        label: project.name,
-        scopeType: "PROJECT",
-        scopeId: project.id,
-      });
-    }
-    return out;
-  }, [
-    organizationId,
-    organizationName,
-    teamId,
-    teamName,
-    projectId,
-    projectName,
-    availableTeams,
-    availableProjects,
-  ]);
-
-  const collection = useMemo(
-    () => createListCollection({ items: options }),
-    [options],
-  );
-
-  const selectedValues = useMemo(
-    () => state.scopes.map((s) => `${s.scopeType}:${s.scopeId}`),
-    [state.scopes],
-  );
 
   if (isExisting) {
     const storedScopes: ScopeSelection[] =
@@ -209,15 +122,9 @@ export function ProviderScopeSection({
 
   if (!hasOrgOrTeam) return null;
 
-  // One-click pre-fills for the common single-scope picks. Each chip
-  // replaces the current selection with exactly that scope so users don't
-  // have to dig into the multi-select for the 90% case ("this project",
-  // "this team", "the whole org"). The multi-select below is still there
-  // for cross-team / cross-project setups.
   const quickPicks = (
     [
       organizationId && {
-        key: "ORGANIZATION" as const,
         label: "Organization",
         icon: <Building2 size={14} aria-hidden />,
         scope: {
@@ -226,20 +133,17 @@ export function ProviderScopeSection({
         },
       },
       teamId && {
-        key: "TEAM" as const,
         label: "This team",
         icon: <Users size={14} aria-hidden />,
         scope: { scopeType: "TEAM" as const, scopeId: teamId },
       },
       projectId && {
-        key: "PROJECT" as const,
         label: "This project",
         icon: <Folder size={14} aria-hidden />,
         scope: { scopeType: "PROJECT" as const, scopeId: projectId },
       },
     ] as const
   ).filter(Boolean) as Array<{
-    key: ModelProviderScopeType;
     label: string;
     icon: React.ReactElement;
     scope: ScopeSelection;
@@ -275,76 +179,19 @@ export function ProviderScopeSection({
           })}
         </Wrap>
       )}
-      <Select.Root
-        collection={collection}
-        value={selectedValues}
-        multiple
-        onValueChange={(details) => {
-          const picked = new Set(details.value);
-          const next = options
-            .filter((o) => picked.has(o.value))
-            .map((o) => ({ scopeType: o.scopeType, scopeId: o.scopeId }));
-          actions.setScopes(next);
-        }}
-      >
-        <Select.Trigger>
-          <Select.ValueText placeholder="Pick one or more scopes">
-            {() => {
-              if (state.scopes.length === 0) return "Pick one or more scopes";
-              return <ProviderScopeChips scopes={state.scopes} />;
-            }}
-          </Select.ValueText>
-        </Select.Trigger>
-        <Select.Content>
-          {options.some((o) => o.scopeType === "ORGANIZATION") && (
-            <Select.ItemGroup label="Organization">
-              {options
-                .filter((o) => o.scopeType === "ORGANIZATION")
-                .map((option) => (
-                  <Select.Item key={option.value} item={option}>
-                    <HStack gap={2}>
-                      <ScopeIcon scopeType="ORGANIZATION" />
-                      <Text>{option.label}</Text>
-                    </HStack>
-                  </Select.Item>
-                ))}
-            </Select.ItemGroup>
-          )}
-          {options.some((o) => o.scopeType === "TEAM") && (
-            <Select.ItemGroup label="Teams">
-              {options
-                .filter((o) => o.scopeType === "TEAM")
-                .map((option) => (
-                  <Select.Item key={option.value} item={option}>
-                    <HStack gap={2}>
-                      <ScopeIcon scopeType="TEAM" />
-                      <Text>{option.label}</Text>
-                    </HStack>
-                  </Select.Item>
-                ))}
-            </Select.ItemGroup>
-          )}
-          {options.some((o) => o.scopeType === "PROJECT") && (
-            <Select.ItemGroup label="Projects">
-              {options
-                .filter((o) => o.scopeType === "PROJECT")
-                .map((option) => (
-                  <Select.Item key={option.value} item={option}>
-                    <HStack gap={2}>
-                      <ScopeIcon scopeType="PROJECT" />
-                      <Text>{option.label}</Text>
-                    </HStack>
-                  </Select.Item>
-                ))}
-            </Select.ItemGroup>
-          )}
-        </Select.Content>
-      </Select.Root>
-      <Box>
-        <Text fontSize="xs" color="gray.600">
-          {summariseSelection(state.scopes)}
-        </Text>
-      </Box>
+      <ScopeChipPicker
+        value={state.scopes}
+        onChange={(next) => actions.setScopes(next)}
+        organizationId={organizationId}
+        organizationName={organizationName}
+        teamId={teamId}
+        teamName={teamName}
+        projectId={projectId}
+        projectName={projectName}
+        availableTeams={availableTeams}
+        availableProjects={availableProjects}
+        label=""
+      />
     </VStack>
   );
 }
